minor avoid creating temporary Flow_Ids
While creating initial and final vertices we repeatedly use the 'Initial and
'Final versions of a flow id. Now we prepare these in advance and re-use.

This is primarily meant to clarify the code, but should also slightly
improve the performance (since Flow_Id is a non-trivial controlled type).

* gnat2why/flow/flow-control_flow_graph.adb
(Create_Initial_And_Final_Vertices.Process): make initial and final versions
 of a flow id and reuse them.

Change-Id: Icd3677ba0735fef24fdaf61d11fdf4f8777d400f