P118-034 XFAIL more tests with isolated new GG problems
Pick front end globals in phase 1 is getting confused by in heavily nested
generic package Ada.Numerics.Discrete_Random. This needs a fixed in phase 1,
most likely in the front end xrefs code which AFAIU is not really aware of
instances of generic units.

Non-blocking for the new GG merge, as diff comes from change in heuristics
for a code not-in-SPARK.

Change-Id: Ib481eada852ea7293120ebb95c6c6ef5942668e3