P118-034 better down-projection in phase 1
When adding globals to graph in phase 1 we take either proper or refined
ones (depending on scope), but still need to down-project them anyway. This
happens when adding to the graph, but similarly, we need to up-project
things when folding them.

Also, some fix for protected type uncovered by this change.

Change-Id: I0f14f22c157b483901b200d1bff5aa408465baa2