P118-034 new messages from the GG branch
Thanks to the new GG we now correctly recognize that heap is manipulated by
subprograms out-of-SPARK. As a consequence, __HEAP appears as a new export
of some MAIN subprograms and since they appear to have an effect, we no
longer inhibit warnings about ineffective statements.

This looks like improvement.

* testsuite/gnatprove/tests/P718-001__ubstr_bad/test.out
* testsuite/gnatprove/tests/P718-001__ubstr_good/test.out
Update test output with new warning and checks.

Change-Id: I3a3b84a1cbcf549f2a91311390cc93d078a0a0a8