P118-034 improved detection of terminating generic routines
In old GG we conservatively assumed that if a generic unit is instantiated
with an actual subprogram then each generic routine calls that actual. Now
we precisely track call chains within such a generic, so we precisely detect
which generic routines are non-terminating due to call to a non-terminating
actual.

This patch removes spurious messages related to non-termination.

Credits to Elisa for the investigation.

Change-Id: I2bd8dda37f1b2116e60097a52a0d705557803a71