Q424-021 fix flow analysis of ghost code
Flow analysis of ghost code didn't really follow the SPARK RM. Instead, it
was based on a vague notion of "something in a proof context" which was
attributed to flow graph vertices.

Now the vertices have a single attribute "Is_Assertion" which is set iff
they come from an "assertion expression", as defined by the Ada RM.

This seemingly simple change affects a lot of code, because the API has
changed. Only non-obvious modifications are documented below.

* gnat2why/flow/flow_utility.ads
* gnat2why/flow/flow_utility.adb
(Is_Ghost_Object): refactored from the old "Refers_To_Ghost", which is now
 removed.

* gnat2why/flow/flow-analysis.adb
(Find_Ineffective_Statements): explicitly suppress messages for ghost
 entities and assignments to ghost variables, which previously happened
 because they caused Is_Proof vertex attribute to be set.

* gnat2why/flow/flow-control_flow_graph.adb
(Process_Subprogram_Globals): set Is_Assertion attribute for reads of
 Proof_In globals.

* gnat2why/flow/flow-control_flow_graph-utility.adb
(Make_Global_Attribute): set Is_Assertion attribute (required on Proof_In
 globals).

* gnat2why/flow/flow-slice.adb
(Compute_Globals): detect proof-in globals and calls based on the
 Is_Assertion vertex attribute; don't rely on the vertex-scanning which is
 done to detect Direct_Calls, because it seems better to do everything
 related to slicing here.

* testsuite/gnatprove/tests/Q424-021__flow_refined_deps/test.out
Update of the primary test for this TN.

* testsuite/gnatprove/tests/P413-011__elaboration_issue_init/test.out
* testsuite/gnatprove/tests/P413-011__elaboration_issue_init2/test.out
* testsuite/gnatprove/tests/P413-011__elaboration_no_issue_init/test.out
* testsuite/gnatprove/tests/PC21-009__flow_assert_initializes/test.out
Better location of messages (now on the first use of offending variable);
not intended but welcomed side-effect.

* testsuite/gnatprove/tests/sparksmt__lexer/lexer.ads
* testsuite/gnatprove/tests/sparksmt__lexer_termination/lexer.ads
* testsuite/gnatprove/tests/text_io_get_line/textio.ads
Fix wrong annotations that would now result in error (a ghost function whose
result depends on a global variable has this variable as Input, not as a
Proof_In).

Change-Id: Id28ba978d616700d90ade0652caeb27785324739